refactor: split slack/discord/session maintenance helpers
(cherry picked from commit 7eda6323249779689ec680c715cfc37434f796fd)
